fixed an issue where tanker aircraft radio comm frequency can not be assigned to service creation dialog. closes #15

+++ src/main/java/moe/ofs/addon/aarservice/domains/Comm.java
@@ -13,7 +13,7 @@ import java.util.Map;
 @Builder
 public class Comm implements Serializable {
     private Map<Object, Object> callsign;
-    private long frequency;
+    private double frequency;
     private int channel;
     private TacanModeChannel modeChannel;
     private String beaconMorseCode;
